+++ collector.h
@@ -24,8 +24,6 @@
 #include "server.h"
 #include "relay.h"
 
-extern int collector_interval;
-
 #define timediff(X, Y) \
 	(Y.tv_sec > X.tv_sec ? (Y.tv_sec - X.tv_sec) * 1000 * 1000 + ((Y.tv_usec - X.tv_usec)) : Y.tv_usec - X.tv_usec)