+++ conffile.tab.c
@@ -228,6 +228,8 @@ union ROUTER_YYSTYPE
   char * statistics_opt_prefix;
   /* cluster  */
   cluster * cluster;
+  /* statistics_opt_counters  */
+  col_mode statistics_opt_counters;
   /* match_send_to  */
   destinations * match_send_to;
   /* match_dsts  */
@@ -262,8 +264,6 @@ union ROUTER_YYSTYPE
   int match_opt_stop;
   /* statistics_opt_interval  */
   int statistics_opt_interval;
-  /* statistics_opt_counters  */
-  int statistics_opt_counters;
   /* cluster_opt_proto  */
   serv_ctype cluster_opt_proto;
   /* aggregate_comp_type  */
@@ -572,9 +572,9 @@ union yyalloc
 #endif /* !YYCOPY_NEEDED */
 
 /* YYFINAL -- State number of the termination state.  */
-#define YYFINAL  30
+#define YYFINAL  31
 /* YYLAST -- Last index in YYTABLE.  */
-#define YYLAST   109
+#define YYLAST   101
 
 /* YYNTOKENS -- Number of terminals.  */
 #define YYNTOKENS  65
@@ -583,7 +583,7 @@ union yyalloc
 /* YYNRULES -- Number of rules.  */
 #define YYNRULES  92
 /* YYNSTATES -- Number of states.  */
-#define YYNSTATES  144
+#define YYNSTATES  145
 
 /* YYTRANSLATE[YYX] -- Symbol number corresponding to YYX as returned
    by yylex, with out-of-bounds checking.  */
@@ -635,16 +635,16 @@ static const yytype_uint8 yytranslate[] =
   /* YYRLINE[YYN] -- Source line where rule number YYN was defined.  */
 static const yytype_uint16 yyrline[] =
 {
-       0,    89,    89,    92,    93,    96,    99,   100,   101,   102,
-     103,   104,   105,   109,   168,   208,   210,   214,   215,   216,
-     219,   220,   223,   224,   225,   228,   229,   232,   233,   236,
-     238,   239,   241,   260,   262,   263,   265,   282,   283,   285,
-     286,   287,   292,   341,   353,   356,   358,   359,   362,   380,
-     381,   399,   400,   403,   406,   416,   419,   421,   422,   425,
-     441,   442,   447,   492,   582,   583,   588,   589,   590,   593,
-     597,   598,   600,   615,   616,   617,   618,   619,   620,   621,
-     631,   632,   635,   636,   641,   653,   661,   670,   671,   682,
-     683,   686,   687
+       0,    90,    90,    93,    94,    97,   100,   101,   102,   103,
+     104,   105,   106,   110,   169,   209,   211,   215,   216,   217,
+     220,   221,   224,   225,   226,   229,   230,   233,   234,   237,
+     239,   240,   242,   261,   263,   264,   266,   283,   284,   286,
+     287,   288,   293,   342,   354,   357,   359,   360,   363,   381,
+     382,   400,   401,   404,   407,   417,   420,   422,   423,   426,
+     442,   443,   448,   493,   583,   584,   589,   590,   591,   594,
+     598,   599,   601,   616,   617,   618,   619,   620,   621,   622,
+     632,   633,   636,   637,   642,   657,   677,   678,   689,   690,
+     693,   694,   699
 };
 #endif
 
@@ -674,9 +674,8 @@ static const char *const yytname[] =
   "match_opt_dst", "match_dst", "match_opt_stop", "rewrite", "aggregate",
   "aggregate_opt_timestamp", "aggregate_ts_when", "aggregate_computes",
   "aggregate_opt_compute", "aggregate_compute", "aggregate_comp_type",
-  "aggregate_opt_send_to", "send", "include", "statistics",
-  "statistics_opt_interval", "statistics_opt_counters",
-  "statistics_opt_prefix", YY_NULLPTR
+  "aggregate_opt_send_to", "send", "statistics", "statistics_opt_interval",
+  "statistics_opt_counters", "statistics_opt_prefix", "include", YY_NULLPTR
 };
 #endif
 
@@ -695,12 +694,12 @@ static const yytype_uint16 yytoknum[] =
 };
 # endif
 
-#define YYPACT_NINF -75
+#define YYPACT_NINF -72
 
 #define yypact_value_is_default(Yystate) \
-  (!!((Yystate) == (-75)))
+  (!!((Yystate) == (-72)))
 
-#define YYTABLE_NINF -4
+#define YYTABLE_NINF -1
 
 #define yytable_value_is_error(Yytable_value) \
   0
@@ -709,21 +708,21 @@ static const yytype_uint16 yytoknum[] =
      STATE-NUM.  */
 static const yytype_int8 yypact[] =
 {
-       0,   -52,   -40,   -32,   -39,   -40,    -4,   -26,    35,     1,
-     -25,   -75,   -75,   -75,   -75,   -75,   -75,   -75,   -17,     6,
-     -75,   -75,    30,   -75,   -11,    27,    24,    25,    -6,   -75,
-     -75,   -75,   -75,   -75,     2,     4,   -75,   -75,   -75,   -75,
-     -75,   -75,    46,     3,    48,    52,     5,     7,    44,   -75,
-     -75,   -22,     8,     9,    39,    40,    14,    44,   -75,    11,
-     -75,     3,   -75,   -75,    12,   -75,   -75,   -75,     5,    56,
-      53,    54,   -75,   -75,    54,   -75,    18,   -75,    47,   -75,
-      26,    20,   -75,    54,    22,    70,   -75,   -75,   -75,   -75,
-     -75,    10,   -22,   -75,   -75,   -75,   -75,   -75,    55,   -75,
-     -75,   -75,   -75,    17,   -75,   -75,   -75,   -75,   -75,    57,
-     -75,   -75,    23,    58,    59,    51,    50,   -30,    -2,    44,
-      50,   -75,   -75,   -75,    49,   -75,   -75,   -75,   -75,   -75,
-     -75,   -75,   -75,   -75,    41,    54,   -75,   -75,    60,    71,
-     -75,   -75,    33,   -75
+      -2,   -37,   -48,   -26,   -33,   -48,   -16,   -23,    39,   -72,
+      -2,   -21,   -72,   -72,   -72,   -72,   -72,   -72,   -72,    25,
+     -72,   -72,    22,   -72,   -17,    19,    17,    16,    18,    -5,
+     -72,   -72,   -72,   -72,   -72,   -72,   -72,   -72,   -72,   -72,
+      35,   -10,    37,    40,    -6,    -3,    31,   -72,   -72,   -22,
+      -1,     0,     1,     4,     7,   -72,    -8,   -72,   -10,   -72,
+     -72,     3,   -72,   -72,   -72,    -6,    41,    43,    42,   -72,
+     -72,    42,   -72,     8,   -72,    44,    45,    38,    12,    31,
+      11,    59,   -72,   -72,   -72,   -72,   -72,    -7,   -22,   -72,
+     -72,   -72,   -72,   -72,    46,   -72,    21,    13,   -72,    42,
+     -72,     6,   -72,   -72,   -72,   -72,   -72,    47,   -72,   -72,
+     -72,   -72,   -72,    20,    48,    49,    50,    51,   -18,   -38,
+      31,    51,   -72,   -72,   -72,    52,   -72,   -72,   -72,   -72,
+     -72,   -72,   -72,   -72,   -72,    30,    42,   -72,   -72,    53,
+      60,   -72,   -72,    26,   -72
 };
 
   /* YYDEFACT[STATE-NUM] -- Default reduction number in state STATE-NUM.
@@ -731,95 +730,95 @@ static const yytype_int8 yypact[] =
      means the default is an error.  */
 static const yytype_uint8 yydefact[] =
 {
-      87,     0,     0,     0,     0,     0,     0,     0,     0,    87,
-       0,     6,     7,     8,     9,    10,    12,    11,    89,     0,
-      48,    43,    49,    44,    46,     0,     0,     0,     0,    85,
-       1,     4,     2,     5,     0,    91,    17,    18,    19,    22,
-      23,    24,    28,     0,    20,    25,     0,     0,     0,    47,
-      45,     0,     0,     0,     0,     0,     0,    82,    27,    37,
-      13,    34,    21,    15,     0,    16,    32,    14,    30,     0,
-       0,    60,    54,    59,    60,    55,    57,    62,     0,    88,
-       0,     0,    83,    60,     0,    39,    35,    33,    26,    29,
-      31,     0,     0,    61,    42,    84,    58,    56,     0,    90,
-      92,    86,    38,     0,    36,    51,    52,    50,    53,     0,
-      40,    41,     0,     0,    64,     0,     0,     0,     0,    82,
-      70,    66,    67,    68,     0,    73,    74,    75,    76,    77,
-      78,    80,    81,    79,     0,    60,    71,    69,     0,     0,
-      63,    65,     0,    72
+       3,     0,     0,     0,     0,     0,    86,     0,     0,     2,
+       3,     0,     6,     7,     8,     9,    10,    11,    12,     0,
+      48,    43,    49,    44,    46,     0,     0,     0,     0,    88,
+      92,     1,     4,     5,    17,    18,    19,    22,    23,    24,
+      28,     0,    20,    25,     0,     0,     0,    47,    45,     0,
+       0,     0,     0,     0,    90,    27,    37,    13,    34,    21,
+      15,     0,    16,    32,    14,    30,     0,     0,    60,    54,
+      59,    60,    55,    57,    62,     0,     0,     0,     0,    82,
+       0,    39,    35,    33,    26,    29,    31,     0,     0,    61,
+      42,    84,    58,    56,     0,    87,     0,     0,    83,    60,
+      38,     0,    36,    51,    52,    50,    53,     0,    89,    91,
+      85,    40,    41,     0,     0,    64,     0,     0,     0,     0,
+      82,    70,    66,    67,    68,     0,    73,    74,    75,    76,
+      77,    78,    80,    81,    79,     0,    60,    71,    69,     0,
+       0,    63,    65,     0,    72
 };
 
   /* YYPGOTO[NTERM-NUM].  */
 static const yytype_int8 yypgoto[] =
 {
-     -75,    86,   -75,   -75,   -75,   -75,   -75,   -75,   -75,   -75,
-     -75,   -75,   -75,   -75,    28,    36,   -75,   -75,   -75,   -75,
-     -75,    93,    76,   -75,   -75,   -75,   -75,    61,    13,    31,
-     -75,   -75,   -74,   -75,   -75,   -75,   -75,   -19,   -75,   -75,
-     -75,   -16,   -75,   -75,   -75,   -75,   -75,   -75
+     -72,   -72,    75,   -72,   -72,   -72,   -72,   -72,   -72,   -72,
+     -72,   -72,   -72,   -72,    23,    29,   -72,   -72,   -72,   -72,
+     -72,    84,    69,   -72,   -72,   -72,   -72,    54,     9,    28,
+     -72,   -72,   -71,   -72,   -72,   -72,   -72,   -27,   -72,   -72,
+     -72,   -25,   -72,   -72,   -72,   -72,   -72,   -72
 };
 
   /* YYDEFGOTO[NTERM-NUM].  */
 static const yytype_int16 yydefgoto[] =
 {
-      -1,     8,    32,     9,    10,    11,    43,    44,    63,    45,
-      65,    46,    67,    89,    68,    60,    87,    61,    85,   104,
-      12,    22,    23,    50,    24,    48,   107,    82,    74,    75,
-      97,    76,    94,    13,    14,   116,   124,   119,   137,   120,
-     134,    83,    15,    16,    17,    18,    35,    57
+      -1,     8,     9,    10,    11,    12,    41,    42,    60,    43,
+      62,    44,    64,    85,    65,    57,    83,    58,    81,   102,
+      13,    22,    23,    48,    24,    46,   105,    98,    71,    72,
+      93,    73,    90,    14,    15,   117,   125,   120,   138,   121,
+     135,    99,    16,    17,    29,    54,    79,    18
 };
 
   /* YYTABLE[YYPACT[STATE-NUM]] -- What to do in state STATE-NUM.  If
      positive, shift that token.  If negative, reduce the rule whose
      number is the opposite.  If YYTABLE_NINF, syntax error.  */
-static const yytype_int16 yytable[] =
+static const yytype_uint8 yytable[] =
 {
-      95,    -3,    72,     1,     1,   121,   122,   123,    19,   101,
-      36,    37,    38,    39,    40,    41,    42,     2,     2,    25,
-      20,    26,     3,     3,    21,    28,     4,     4,     5,     5,
-     105,   106,   110,   111,    29,    30,    34,    33,    73,   125,
-     126,   127,   128,   129,   130,   131,   132,   133,    47,    20,
-      51,    52,     6,     6,    53,    54,    55,    58,     7,     7,
-      56,   140,    62,    59,    64,    66,    70,    69,    77,    79,
-      78,    81,    80,    88,    84,    91,    92,    98,    73,    93,
-     100,    99,   102,   103,   113,   117,   109,   138,   114,   112,
-     118,   139,   115,   143,   142,    31,    90,    86,    27,   141,
-      49,   136,     0,   135,     0,   108,     0,    96,     0,    71
+      91,     1,    69,   126,   127,   128,   129,   130,   131,   132,
+     133,   134,    20,   103,   104,     2,    21,   122,   123,   124,
+       3,   111,   112,    19,     4,    25,     5,    26,   110,    34,
+      35,    36,    37,    38,    39,    40,    28,    30,    70,    31,
+      45,    33,    49,    20,    50,    51,    55,    52,    53,     6,
+      56,    59,    61,    67,    63,    80,     7,    66,    77,    74,
+      87,    75,    76,    78,    84,   141,    88,    89,    70,    97,
+      96,   100,   101,   109,    94,    95,   108,   107,   115,   113,
+     140,   114,   116,   143,   118,    32,   144,    82,    86,    27,
+     139,   119,   142,    47,   137,   136,     0,   106,     0,     0,
+      68,    92
 };
 
 static const yytype_int16 yycheck[] =
 {
-      74,     0,    24,     3,     3,    35,    36,    37,    60,    83,
-       4,     5,     6,     7,     8,     9,    10,    17,    17,    51,
-      60,    60,    22,    22,    64,    29,    26,    26,    28,    28,
-      20,    21,    15,    16,    60,     0,    53,    62,    60,    41,
-      42,    43,    44,    45,    46,    47,    48,    49,    18,    60,
-      23,    27,    52,    52,    29,    61,    54,    11,    58,    58,
-      56,   135,    14,    60,    12,    60,    22,    60,    60,    30,
-      61,    57,    32,    61,    63,    19,    23,    30,    60,    25,
-      60,    55,    60,    13,    61,    34,    31,    38,    30,    32,
-      40,    50,    33,    60,    23,     9,    68,    61,     5,    39,
-      24,   120,    -1,   119,    -1,    92,    -1,    76,    -1,    48
+      71,     3,    24,    41,    42,    43,    44,    45,    46,    47,
+      48,    49,    60,    20,    21,    17,    64,    35,    36,    37,
+      22,    15,    16,    60,    26,    51,    28,    60,    99,     4,
+       5,     6,     7,     8,     9,    10,    52,    60,    60,     0,
+      18,    62,    23,    60,    27,    29,    11,    29,    53,    51,
+      60,    14,    12,    22,    60,    63,    58,    60,    54,    60,
+      19,    61,    61,    56,    61,   136,    23,    25,    60,    57,
+      32,    60,    13,    60,    30,    30,    55,    31,    30,    32,
+      50,    61,    33,    23,    34,    10,    60,    58,    65,     5,
+      38,    40,    39,    24,   121,   120,    -1,    88,    -1,    -1,
+      46,    73
 };
 
   /* YYSTOS[STATE-NUM] -- The (internal number of the) accessing
      symbol of state STATE-NUM.  */
 static const yytype_uint8 yystos[] =
 {
-       0,     3,    17,    22,    26,    28,    52,    58,    66,    68,
-      69,    70,    85,    98,    99,   107,   108,   109,   110,    60,
-      60,    64,    86,    87,    89,    51,    60,    86,    29,    60,
-       0,    66,    67,    62,    53,   111,     4,     5,     6,     7,
-       8,     9,    10,    71,    72,    74,    76,    18,    90,    87,
-      88,    23,    27,    29,    61,    54,    56,   112,    11,    60,
-      80,    82,    14,    73,    12,    75,    60,    77,    79,    60,
-      22,    92,    24,    60,    93,    94,    96,    60,    61,    30,
-      32,    57,    92,   106,    63,    83,    80,    81,    61,    78,
-      79,    19,    23,    25,    97,    97,    94,    95,    30,    55,
-      60,    97,    60,    13,    84,    20,    21,    91,    93,    31,
-      15,    16,    32,    61,    30,    33,   100,    34,    40,   102,
-     104,    35,    36,    37,   101,    41,    42,    43,    44,    45,
-      46,    47,    48,    49,   105,   106,   102,   103,    38,    50,
-      97,    39,    23,    60
+       0,     3,    17,    22,    26,    28,    51,    58,    66,    67,
+      68,    69,    70,    85,    98,    99,   107,   108,   112,    60,
+      60,    64,    86,    87,    89,    51,    60,    86,    52,   109,
+      60,     0,    67,    62,     4,     5,     6,     7,     8,     9,
+      10,    71,    72,    74,    76,    18,    90,    87,    88,    23,
+      27,    29,    29,    53,   110,    11,    60,    80,    82,    14,
+      73,    12,    75,    60,    77,    79,    60,    22,    92,    24,
+      60,    93,    94,    96,    60,    61,    61,    54,    56,   111,
+      63,    83,    80,    81,    61,    78,    79,    19,    23,    25,
+      97,    97,    94,    95,    30,    30,    32,    57,    92,   106,
+      60,    13,    84,    20,    21,    91,    93,    31,    55,    60,
+      97,    15,    16,    32,    61,    30,    33,   100,    34,    40,
+     102,   104,    35,    36,    37,   101,    41,    42,    43,    44,
+      45,    46,    47,    48,    49,   105,   106,   102,   103,    38,
+      50,    97,    39,    23,    60
 };
 
   /* YYR1[YYN] -- Symbol number of symbol that rule YYN derives.  */
@@ -833,14 +832,14 @@ static const yytype_uint8 yyr1[] =
       90,    91,    91,    92,    93,    93,    94,    95,    95,    96,
       97,    97,    98,    99,   100,   100,   101,   101,   101,   102,
      103,   103,   104,   105,   105,   105,   105,   105,   105,   105,
-     105,   105,   106,   106,   107,   108,   109,   110,   110,   111,
-     111,   112,   112
+     105,   105,   106,   106,   107,   108,   109,   109,   110,   110,
+     111,   111,   112
 };
 
   /* YYR2[YYN] -- Number of symbols on the right hand side of rule YYN.  */
 static const yytype_uint8 yyr2[] =
 {
-       0,     2,     2,     0,     1,     2,     1,     1,     1,     1,
+       0,     2,     1,     0,     2,     2,     1,     1,     1,     1,
        1,     1,     1,     4,     4,     2,     2,     1,     1,     1,
        0,     1,     1,     1,     1,     0,     2,     2,     1,     2,
        0,     1,     1,     2,     0,     1,     3,     0,     2,     0,
@@ -848,8 +847,8 @@ static const yytype_uint8 yyr2[] =
        4,     1,     1,     3,     1,     1,     2,     0,     1,     1,
        0,     1,     4,    13,     0,     5,     1,     1,     1,     2,
        0,     1,     5,     1,     1,     1,     1,     1,     1,     1,
-       1,     1,     0,     1,     5,     2,     5,     0,     4,     0,
-       4,     0,     3
+       1,     1,     0,     1,     5,     6,     0,     4,     0,     4,
+       0,     3,     2
 };
 
 
@@ -1629,7 +1628,7 @@ yyreduce:
   switch (yyn)
     {
         case 13:
-#line 110 "conffile.y" /* yacc.c:1646  */
+#line 111 "conffile.y" /* yacc.c:1646  */
     {
 	   	struct _clhost *w;
 		char *err;
@@ -1688,11 +1687,11 @@ yyreduce:
 			YYERROR;
 		}
 	   }
-#line 1692 "conffile.tab.c" /* yacc.c:1646  */
+#line 1691 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 14:
-#line 169 "conffile.y" /* yacc.c:1646  */
+#line 170 "conffile.y" /* yacc.c:1646  */
     {
 	   	struct _clhost *w;
 		char *err;
@@ -1729,113 +1728,113 @@ yyreduce:
 			YYERROR;
 		}
 	   }
-#line 1733 "conffile.tab.c" /* yacc.c:1646  */
+#line 1732 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 15:
-#line 209 "conffile.y" /* yacc.c:1646  */
+#line 210 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clust*)(&yyval)).t = (*(enum clusttype*)(&yyvsp[-1])); (*(struct _clust*)(&yyval)).ival = (*(int*)(&yyvsp[0])); }
-#line 1739 "conffile.tab.c" /* yacc.c:1646  */
+#line 1738 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 16:
-#line 211 "conffile.y" /* yacc.c:1646  */
+#line 212 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clust*)(&yyval)).t = (*(enum clusttype*)(&yyvsp[-1])); (*(struct _clust*)(&yyval)).ival = (*(int*)(&yyvsp[0])); }
-#line 1745 "conffile.tab.c" /* yacc.c:1646  */
+#line 1744 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 17:
-#line 214 "conffile.y" /* yacc.c:1646  */
+#line 215 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = FORWARD; }
-#line 1751 "conffile.tab.c" /* yacc.c:1646  */
+#line 1750 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 18:
-#line 215 "conffile.y" /* yacc.c:1646  */
+#line 216 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = ANYOF; }
-#line 1757 "conffile.tab.c" /* yacc.c:1646  */
+#line 1756 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 19:
-#line 216 "conffile.y" /* yacc.c:1646  */
+#line 217 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = FAILOVER; }
-#line 1763 "conffile.tab.c" /* yacc.c:1646  */
+#line 1762 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 20:
-#line 219 "conffile.y" /* yacc.c:1646  */
+#line 220 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 0; }
-#line 1769 "conffile.tab.c" /* yacc.c:1646  */
+#line 1768 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 21:
-#line 220 "conffile.y" /* yacc.c:1646  */
+#line 221 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 1; }
-#line 1775 "conffile.tab.c" /* yacc.c:1646  */
+#line 1774 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 22:
-#line 223 "conffile.y" /* yacc.c:1646  */
+#line 224 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = CARBON_CH; }
-#line 1781 "conffile.tab.c" /* yacc.c:1646  */
+#line 1780 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 23:
-#line 224 "conffile.y" /* yacc.c:1646  */
+#line 225 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = FNV1A_CH; }
-#line 1787 "conffile.tab.c" /* yacc.c:1646  */
+#line 1786 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 24:
-#line 225 "conffile.y" /* yacc.c:1646  */
+#line 226 "conffile.y" /* yacc.c:1646  */
     { (*(enum clusttype*)(&yyval)) = JUMP_CH; }
-#line 1793 "conffile.tab.c" /* yacc.c:1646  */
+#line 1792 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 25:
-#line 228 "conffile.y" /* yacc.c:1646  */
+#line 229 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 1; }
-#line 1799 "conffile.tab.c" /* yacc.c:1646  */
+#line 1798 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 26:
-#line 229 "conffile.y" /* yacc.c:1646  */
+#line 230 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = (*(int*)(&yyvsp[0])); }
-#line 1805 "conffile.tab.c" /* yacc.c:1646  */
+#line 1804 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 27:
-#line 232 "conffile.y" /* yacc.c:1646  */
+#line 233 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clust*)(&yyval)).t = FILELOGIP; (*(struct _clust*)(&yyval)).ival = 0; }
-#line 1811 "conffile.tab.c" /* yacc.c:1646  */
+#line 1810 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 28:
-#line 233 "conffile.y" /* yacc.c:1646  */
+#line 234 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clust*)(&yyval)).t = FILELOG; (*(struct _clust*)(&yyval)).ival = 0; }
-#line 1817 "conffile.tab.c" /* yacc.c:1646  */
+#line 1816 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 29:
-#line 236 "conffile.y" /* yacc.c:1646  */
+#line 237 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyvsp[-1]))->next = (*(struct _clhost **)(&yyvsp[0])); (*(struct _clhost **)(&yyval)) = (*(struct _clhost **)(&yyvsp[-1])); }
-#line 1823 "conffile.tab.c" /* yacc.c:1646  */
+#line 1822 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 30:
-#line 238 "conffile.y" /* yacc.c:1646  */
+#line 239 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyval)) = NULL; }
-#line 1829 "conffile.tab.c" /* yacc.c:1646  */
+#line 1828 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 31:
-#line 239 "conffile.y" /* yacc.c:1646  */
+#line 240 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyval)) = (*(struct _clhost **)(&yyvsp[0])); }
-#line 1835 "conffile.tab.c" /* yacc.c:1646  */
+#line 1834 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 32:
-#line 242 "conffile.y" /* yacc.c:1646  */
+#line 243 "conffile.y" /* yacc.c:1646  */
     {
 				struct _clhost *ret = ra_malloc(rtr, sizeof(struct _clhost));
 				char *err = router_validate_path(rtr, (*(char **)(&yyvsp[0])));
@@ -1852,29 +1851,29 @@ yyreduce:
 				ret->next = NULL;
 				(*(struct _clhost **)(&yyval)) = ret;
 			}
-#line 1856 "conffile.tab.c" /* yacc.c:1646  */
+#line 1855 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 33:
-#line 260 "conffile.y" /* yacc.c:1646  */
+#line 261 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyvsp[-1]))->next = (*(struct _clhost **)(&yyvsp[0])); (*(struct _clhost **)(&yyval)) = (*(struct _clhost **)(&yyvsp[-1])); }
-#line 1862 "conffile.tab.c" /* yacc.c:1646  */
+#line 1861 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 34:
-#line 262 "conffile.y" /* yacc.c:1646  */
+#line 263 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyval)) = NULL; }
-#line 1868 "conffile.tab.c" /* yacc.c:1646  */
+#line 1867 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 35:
-#line 263 "conffile.y" /* yacc.c:1646  */
+#line 264 "conffile.y" /* yacc.c:1646  */
     { (*(struct _clhost **)(&yyval)) = (*(struct _clhost **)(&yyvsp[0])); }
-#line 1874 "conffile.tab.c" /* yacc.c:1646  */
+#line 1873 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 36:
-#line 266 "conffile.y" /* yacc.c:1646  */
+#line 267 "conffile.y" /* yacc.c:1646  */
     {
 			  	struct _clhost *ret = ra_malloc(rtr, sizeof(struct _clhost));
 				char *err = router_validate_address(
@@ -1890,41 +1889,41 @@ yyreduce:
 				ret->next = NULL;
 				(*(struct _clhost **)(&yyval)) = ret;
 			  }
-#line 1894 "conffile.tab.c" /* yacc.c:1646  */
+#line 1893 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 37:
-#line 282 "conffile.y" /* yacc.c:1646  */
+#line 283 "conffile.y" /* yacc.c:1646  */
     { (*(char **)(&yyval)) = NULL; }
-#line 1900 "conffile.tab.c" /* yacc.c:1646  */
+#line 1899 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 38:
-#line 283 "conffile.y" /* yacc.c:1646  */
+#line 284 "conffile.y" /* yacc.c:1646  */
     { (*(char **)(&yyval)) = (*(char **)(&yyvsp[0])); }
-#line 1906 "conffile.tab.c" /* yacc.c:1646  */
+#line 1905 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 39:
-#line 285 "conffile.y" /* yacc.c:1646  */
+#line 286 "conffile.y" /* yacc.c:1646  */
     { (*(serv_ctype*)(&yyval)) = CON_TCP; }
-#line 1912 "conffile.tab.c" /* yacc.c:1646  */
+#line 1911 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 40:
-#line 286 "conffile.y" /* yacc.c:1646  */
+#line 287 "conffile.y" /* yacc.c:1646  */
     { (*(serv_ctype*)(&yyval)) = CON_UDP; }
-#line 1918 "conffile.tab.c" /* yacc.c:1646  */
+#line 1917 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 41:
-#line 287 "conffile.y" /* yacc.c:1646  */
+#line 288 "conffile.y" /* yacc.c:1646  */
     { (*(serv_ctype*)(&yyval)) = CON_TCP; }
-#line 1924 "conffile.tab.c" /* yacc.c:1646  */
+#line 1923 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 42:
-#line 294 "conffile.y" /* yacc.c:1646  */
+#line 295 "conffile.y" /* yacc.c:1646  */
     {
 	 	/* each expr comes with an allocated route, populate it */
 		struct _maexpr *we;
@@ -1970,11 +1969,11 @@ yyreduce:
 			}
 		}
 	 }
-#line 1974 "conffile.tab.c" /* yacc.c:1646  */
+#line 1973 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 43:
-#line 342 "conffile.y" /* yacc.c:1646  */
+#line 343 "conffile.y" /* yacc.c:1646  */
     {
 			if (((*(struct _maexpr **)(&yyval)) = ra_malloc(rtr, sizeof(struct _maexpr))) == NULL) {
 				logerr("out of memory\n");
@@ -1986,35 +1985,35 @@ yyreduce:
 			(*(struct _maexpr **)(&yyval))->drop = 0;
 			(*(struct _maexpr **)(&yyval))->next = NULL;
 		   }
-#line 1990 "conffile.tab.c" /* yacc.c:1646  */
+#line 1989 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 44:
-#line 353 "conffile.y" /* yacc.c:1646  */
+#line 354 "conffile.y" /* yacc.c:1646  */
     { (*(struct _maexpr **)(&yyval)) = (*(struct _maexpr **)(&yyvsp[0])); }
-#line 1996 "conffile.tab.c" /* yacc.c:1646  */
+#line 1995 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 45:
-#line 356 "conffile.y" /* yacc.c:1646  */
+#line 357 "conffile.y" /* yacc.c:1646  */
     { (*(struct _maexpr **)(&yyvsp[-1]))->next = (*(struct _maexpr **)(&yyvsp[0])); (*(struct _maexpr **)(&yyval)) = (*(struct _maexpr **)(&yyvsp[-1])); }
-#line 2002 "conffile.tab.c" /* yacc.c:1646  */
+#line 2001 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 46:
-#line 358 "conffile.y" /* yacc.c:1646  */
+#line 359 "conffile.y" /* yacc.c:1646  */
     { (*(struct _maexpr **)(&yyval)) = NULL; }
-#line 2008 "conffile.tab.c" /* yacc.c:1646  */
+#line 2007 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 47:
-#line 359 "conffile.y" /* yacc.c:1646  */
+#line 360 "conffile.y" /* yacc.c:1646  */
     { (*(struct _maexpr **)(&yyval)) = (*(struct _maexpr **)(&yyvsp[0])); }
-#line 2014 "conffile.tab.c" /* yacc.c:1646  */
+#line 2013 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 48:
-#line 363 "conffile.y" /* yacc.c:1646  */
+#line 364 "conffile.y" /* yacc.c:1646  */
     {
 			char *err;
 			if (((*(struct _maexpr **)(&yyval)) = ra_malloc(rtr, sizeof(struct _maexpr))) == NULL) {
@@ -2030,17 +2029,17 @@ yyreduce:
 			(*(struct _maexpr **)(&yyval))->drop = 0;
 			(*(struct _maexpr **)(&yyval))->next = NULL;
 		  }
-#line 2034 "conffile.tab.c" /* yacc.c:1646  */
+#line 2033 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 49:
-#line 380 "conffile.y" /* yacc.c:1646  */
+#line 381 "conffile.y" /* yacc.c:1646  */
     { (*(struct _maexpr **)(&yyval)) = NULL; }
-#line 2040 "conffile.tab.c" /* yacc.c:1646  */
+#line 2039 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 50:
-#line 382 "conffile.y" /* yacc.c:1646  */
+#line 383 "conffile.y" /* yacc.c:1646  */
     {
 					char *err;
 					if (((*(struct _maexpr **)(&yyval)) = ra_malloc(rtr, sizeof(struct _maexpr))) == NULL) {
@@ -2056,29 +2055,29 @@ yyreduce:
 					(*(struct _maexpr **)(&yyval))->drop = (*(int*)(&yyvsp[0]));
 					(*(struct _maexpr **)(&yyval))->next = NULL;
 				  }
-#line 2060 "conffile.tab.c" /* yacc.c:1646  */
+#line 2059 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 51:
-#line 399 "conffile.y" /* yacc.c:1646  */
+#line 400 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 0; }
-#line 2066 "conffile.tab.c" /* yacc.c:1646  */
+#line 2065 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 52:
-#line 400 "conffile.y" /* yacc.c:1646  */
+#line 401 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 1; }
-#line 2072 "conffile.tab.c" /* yacc.c:1646  */
+#line 2071 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 53:
-#line 403 "conffile.y" /* yacc.c:1646  */
+#line 404 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = (*(destinations **)(&yyvsp[0])); }
-#line 2078 "conffile.tab.c" /* yacc.c:1646  */
+#line 2077 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 54:
-#line 407 "conffile.y" /* yacc.c:1646  */
+#line 408 "conffile.y" /* yacc.c:1646  */
     {
 			if (((*(destinations **)(&yyval)) = ra_malloc(rtr, sizeof(destinations))) == NULL) {
 				logerr("out of memory\n");
@@ -2088,35 +2087,35 @@ yyreduce:
 				YYABORT;
 			(*(destinations **)(&yyval))->next = NULL;
 		  }
-#line 2092 "conffile.tab.c" /* yacc.c:1646  */
+#line 2091 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 55:
-#line 416 "conffile.y" /* yacc.c:1646  */
+#line 417 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = (*(destinations **)(&yyvsp[0])); }
-#line 2098 "conffile.tab.c" /* yacc.c:1646  */
+#line 2097 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 56:
-#line 419 "conffile.y" /* yacc.c:1646  */
+#line 420 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyvsp[-1]))->next = (*(destinations **)(&yyvsp[0])); (*(destinations **)(&yyval)) = (*(destinations **)(&yyvsp[-1])); }
-#line 2104 "conffile.tab.c" /* yacc.c:1646  */
+#line 2103 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 57:
-#line 421 "conffile.y" /* yacc.c:1646  */
+#line 422 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = NULL; }
-#line 2110 "conffile.tab.c" /* yacc.c:1646  */
+#line 2109 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 58:
-#line 422 "conffile.y" /* yacc.c:1646  */
+#line 423 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = (*(destinations **)(&yyvsp[0])); }
-#line 2116 "conffile.tab.c" /* yacc.c:1646  */
+#line 2115 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 59:
-#line 426 "conffile.y" /* yacc.c:1646  */
+#line 427 "conffile.y" /* yacc.c:1646  */
     {
 			char *err;
 			if (((*(destinations **)(&yyval)) = ra_malloc(rtr, sizeof(destinations))) == NULL) {
@@ -2130,23 +2129,23 @@ yyreduce:
 			}
 			(*(destinations **)(&yyval))->next = NULL;
 		 }
-#line 2134 "conffile.tab.c" /* yacc.c:1646  */
+#line 2133 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 60:
-#line 441 "conffile.y" /* yacc.c:1646  */
+#line 442 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 0; }
-#line 2140 "conffile.tab.c" /* yacc.c:1646  */
+#line 2139 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 61:
-#line 442 "conffile.y" /* yacc.c:1646  */
+#line 443 "conffile.y" /* yacc.c:1646  */
     { (*(int*)(&yyval)) = 1; }
-#line 2146 "conffile.tab.c" /* yacc.c:1646  */
+#line 2145 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 62:
-#line 448 "conffile.y" /* yacc.c:1646  */
+#line 449 "conffile.y" /* yacc.c:1646  */
     {
 		char *err;
 		route *r = NULL;
@@ -2187,11 +2186,11 @@ yyreduce:
 			YYERROR;
 		}
 	   }
-#line 2191 "conffile.tab.c" /* yacc.c:1646  */
+#line 2190 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 63:
-#line 498 "conffile.y" /* yacc.c:1646  */
+#line 499 "conffile.y" /* yacc.c:1646  */
     {
 		 	cluster *w;
 			aggregator *a;
@@ -2274,59 +2273,59 @@ yyreduce:
 			if ((*(destinations **)(&yyvsp[-1])) != NULL)
 				router_add_stubroute(rtr, AGGRSTUB, w, (*(destinations **)(&yyvsp[-1])));
 		 }
-#line 2278 "conffile.tab.c" /* yacc.c:1646  */
+#line 2277 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 64:
-#line 582 "conffile.y" /* yacc.c:1646  */
+#line 583 "conffile.y" /* yacc.c:1646  */
     { (*(enum _aggr_timestamp*)(&yyval)) = TS_END; }
-#line 2284 "conffile.tab.c" /* yacc.c:1646  */
+#line 2283 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 65:
-#line 585 "conffile.y" /* yacc.c:1646  */
+#line 586 "conffile.y" /* yacc.c:1646  */
     { (*(enum _aggr_timestamp*)(&yyval)) = (*(enum _aggr_timestamp*)(&yyvsp[-2])); }
-#line 2290 "conffile.tab.c" /* yacc.c:1646  */
+#line 2289 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 66:
-#line 588 "conffile.y" /* yacc.c:1646  */
+#line 589 "conffile.y" /* yacc.c:1646  */
     { (*(enum _aggr_timestamp*)(&yyval)) = TS_START; }
-#line 2296 "conffile.tab.c" /* yacc.c:1646  */
+#line 2295 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 67:
-#line 589 "conffile.y" /* yacc.c:1646  */
+#line 590 "conffile.y" /* yacc.c:1646  */
     { (*(enum _aggr_timestamp*)(&yyval)) = TS_MIDDLE; }
-#line 2302 "conffile.tab.c" /* yacc.c:1646  */
+#line 2301 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 68:
-#line 590 "conffile.y" /* yacc.c:1646  */
+#line 591 "conffile.y" /* yacc.c:1646  */
     { (*(enum _aggr_timestamp*)(&yyval)) = TS_END; }
-#line 2308 "conffile.tab.c" /* yacc.c:1646  */
+#line 2307 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 69:
-#line 594 "conffile.y" /* yacc.c:1646  */
+#line 595 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp **)(&yyvsp[-1]))->next = (*(struct _agcomp **)(&yyvsp[0])); (*(struct _agcomp **)(&yyval)) = (*(struct _agcomp **)(&yyvsp[-1])); }
-#line 2314 "conffile.tab.c" /* yacc.c:1646  */
+#line 2313 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 70:
-#line 597 "conffile.y" /* yacc.c:1646  */
+#line 598 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp **)(&yyval)) = NULL; }
-#line 2320 "conffile.tab.c" /* yacc.c:1646  */
+#line 2319 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 71:
-#line 598 "conffile.y" /* yacc.c:1646  */
+#line 599 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp **)(&yyval)) = (*(struct _agcomp **)(&yyvsp[0])); }
-#line 2326 "conffile.tab.c" /* yacc.c:1646  */
+#line 2325 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 72:
-#line 602 "conffile.y" /* yacc.c:1646  */
+#line 603 "conffile.y" /* yacc.c:1646  */
     {
 					(*(struct _agcomp **)(&yyval)) = ra_malloc(rtr, sizeof(struct _agcomp));
 					if ((*(struct _agcomp **)(&yyval)) == NULL) {
@@ -2338,47 +2337,47 @@ yyreduce:
 					(*(struct _agcomp **)(&yyval))->metric = (*(char **)(&yyvsp[0]));
 					(*(struct _agcomp **)(&yyval))->next = NULL;
 				 }
-#line 2342 "conffile.tab.c" /* yacc.c:1646  */
+#line 2341 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 73:
-#line 615 "conffile.y" /* yacc.c:1646  */
+#line 616 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = SUM; }
-#line 2348 "conffile.tab.c" /* yacc.c:1646  */
+#line 2347 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 74:
-#line 616 "conffile.y" /* yacc.c:1646  */
+#line 617 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = CNT; }
-#line 2354 "conffile.tab.c" /* yacc.c:1646  */
+#line 2353 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 75:
-#line 617 "conffile.y" /* yacc.c:1646  */
+#line 618 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = MAX; }
-#line 2360 "conffile.tab.c" /* yacc.c:1646  */
+#line 2359 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 76:
-#line 618 "conffile.y" /* yacc.c:1646  */
+#line 619 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = MIN; }
-#line 2366 "conffile.tab.c" /* yacc.c:1646  */
+#line 2365 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 77:
-#line 619 "conffile.y" /* yacc.c:1646  */
+#line 620 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = AVG; }
-#line 2372 "conffile.tab.c" /* yacc.c:1646  */
+#line 2371 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 78:
-#line 620 "conffile.y" /* yacc.c:1646  */
+#line 621 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = MEDN; }
-#line 2378 "conffile.tab.c" /* yacc.c:1646  */
+#line 2377 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 79:
-#line 622 "conffile.y" /* yacc.c:1646  */
+#line 623 "conffile.y" /* yacc.c:1646  */
     {
 				    if ((*(int*)(&yyvsp[0])) < 1 || (*(int*)(&yyvsp[0])) > 99) {
 						router_yyerror(&yylloc, yyscanner, rtr,
@@ -2388,68 +2387,73 @@ yyreduce:
 				   	(*(struct _agcomp*)(&yyval)).ctype = PCTL;
 					(*(struct _agcomp*)(&yyval)).pctl = (unsigned char)(*(int*)(&yyvsp[0]));
 				   }
-#line 2392 "conffile.tab.c" /* yacc.c:1646  */
+#line 2391 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 80:
-#line 631 "conffile.y" /* yacc.c:1646  */
+#line 632 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = VAR; }
-#line 2398 "conffile.tab.c" /* yacc.c:1646  */
+#line 2397 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 81:
-#line 632 "conffile.y" /* yacc.c:1646  */
+#line 633 "conffile.y" /* yacc.c:1646  */
     { (*(struct _agcomp*)(&yyval)).ctype = SDEV; }
-#line 2404 "conffile.tab.c" /* yacc.c:1646  */
+#line 2403 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 82:
-#line 635 "conffile.y" /* yacc.c:1646  */
+#line 636 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = NULL; }
-#line 2410 "conffile.tab.c" /* yacc.c:1646  */
+#line 2409 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 83:
-#line 636 "conffile.y" /* yacc.c:1646  */
+#line 637 "conffile.y" /* yacc.c:1646  */
     { (*(destinations **)(&yyval)) = (*(destinations **)(&yyvsp[0])); }
-#line 2416 "conffile.tab.c" /* yacc.c:1646  */
+#line 2415 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 84:
-#line 642 "conffile.y" /* yacc.c:1646  */
+#line 643 "conffile.y" /* yacc.c:1646  */
     {
 		char *err = router_set_statistics(rtr, (*(destinations **)(&yyvsp[-1])));
 		if (err != NULL) {
 			router_yyerror(&yylloc, yyscanner, rtr, err);
 			YYERROR;
 		}
+		logerr("warning: 'send statistics to ...' is deprecated and will be "
+				"removed in a future version, use 'statistics send to ...' "
+				"instead\n");
 	}
-#line 2428 "conffile.tab.c" /* yacc.c:1646  */
+#line 2430 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 85:
-#line 654 "conffile.y" /* yacc.c:1646  */
+#line 663 "conffile.y" /* yacc.c:1646  */
     {
-	   	router_readconfig(rtr, (*(char **)(&yyvsp[0])), 0, 0, 0, 0, 0);
-	   }
-#line 2436 "conffile.tab.c" /* yacc.c:1646  */
+		  	char *err;
+		  	router_setcollectorvals(rtr, (*(int*)(&yyvsp[-4])), (*(char **)(&yyvsp[-2])), (*(col_mode*)(&yyvsp[-3])));
+
+			if ((*(destinations **)(&yyvsp[-1])) != NULL) {
+				err = router_set_statistics(rtr, (*(destinations **)(&yyvsp[-1])));
+				if (err != NULL) {
+					router_yyerror(&yylloc, yyscanner, rtr, err);
+					YYERROR;
+				}
+			}
+		  }
+#line 2447 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 86:
-#line 666 "conffile.y" /* yacc.c:1646  */
-    {
-		  }
-#line 2443 "conffile.tab.c" /* yacc.c:1646  */
+#line 677 "conffile.y" /* yacc.c:1646  */
+    { (*(int*)(&yyval)) = 0; }
+#line 2453 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 87:
-#line 670 "conffile.y" /* yacc.c:1646  */
-    { (*(int*)(&yyval)) = 0; }
-#line 2449 "conffile.tab.c" /* yacc.c:1646  */
-    break;
-
-  case 88:
-#line 672 "conffile.y" /* yacc.c:1646  */
+#line 679 "conffile.y" /* yacc.c:1646  */
     {
 					   	if ((*(int*)(&yyvsp[-1])) <= 0) {
 							router_yyerror(&yylloc, yyscanner, rtr,
@@ -2458,35 +2462,43 @@ yyreduce:
 						}
 						(*(int*)(&yyval)) = (*(int*)(&yyvsp[-1]));
 					   }
-#line 2462 "conffile.tab.c" /* yacc.c:1646  */
+#line 2466 "conffile.tab.c" /* yacc.c:1646  */
+    break;
+
+  case 88:
+#line 689 "conffile.y" /* yacc.c:1646  */
+    { (*(col_mode*)(&yyval)) = CUM; }
+#line 2472 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 89:
-#line 682 "conffile.y" /* yacc.c:1646  */
-    { (*(int*)(&yyval)) = 0; }
-#line 2468 "conffile.tab.c" /* yacc.c:1646  */
+#line 690 "conffile.y" /* yacc.c:1646  */
+    { (*(col_mode*)(&yyval)) = SUB; }
+#line 2478 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 90:
-#line 683 "conffile.y" /* yacc.c:1646  */
-    { (*(int*)(&yyval)) = 1; }
-#line 2474 "conffile.tab.c" /* yacc.c:1646  */
+#line 693 "conffile.y" /* yacc.c:1646  */
+    { (*(char **)(&yyval)) = NULL; }
+#line 2484 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 91:
-#line 686 "conffile.y" /* yacc.c:1646  */
-    { (*(char **)(&yyval)) = NULL; }
-#line 2480 "conffile.tab.c" /* yacc.c:1646  */
+#line 694 "conffile.y" /* yacc.c:1646  */
+    { (*(char **)(&yyval)) = (*(char **)(&yyvsp[0])); }
+#line 2490 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
   case 92:
-#line 687 "conffile.y" /* yacc.c:1646  */
-    { (*(char **)(&yyval)) = (*(char **)(&yyvsp[0])); }
-#line 2486 "conffile.tab.c" /* yacc.c:1646  */
+#line 700 "conffile.y" /* yacc.c:1646  */
+    {
+	   	router_readconfig(rtr, (*(char **)(&yyvsp[0])), 0, 0, 0, 0, 0);
+	   }
+#line 2498 "conffile.tab.c" /* yacc.c:1646  */
     break;
 
 
-#line 2490 "conffile.tab.c" /* yacc.c:1646  */
+#line 2502 "conffile.tab.c" /* yacc.c:1646  */
       default: break;
     }
   /* User semantic actions sometimes alter yychar, and that requires